+++ utilcode/pkg/src/main/java/com/blankj/utilcode/pkg/feature/activity/ActivityActivity.kt
@@ -54,7 +54,6 @@ class ActivityActivity : BaseBackActivity() {
 
 
     override fun initView(savedInstanceState: Bundle?, contentView: View) {
-
         setTitle(R.string.demo_activity)
 
         activityClzBtn.setOnClickListener(this)